Make blackboard toolbar darker; Update release notes; Put Explain's 2-year plan on the home page; Include GitHub in the home page;
