+++ src/pages/AllOpenSpaces.vue
@@ -67,30 +67,47 @@
 
     <portal to="main-content">
       <div class="ma-5">
-        <h2>This week's updates</h2>
+        <h2>New improvements</h2>
         <ul>
-          <li>Everyone can now freely create new classes, open spaces and rooms</li>
+          <li>
+            The blackboard is easier to use now, with a larger effective space, and is scrollable by default. 
+            Switching between different blackboards will no longer reset your pen settings, and the pen strokes are finer to make it easier to write subscripts.
+          </li>
+          <li>
+            The left-drawer has more effective space, so you can see more rooms and participants at once without having to scroll. 
+          </li>
         </ul>
         
         <br>
-        <h2>
-          Tips
-        </h2>
+
+        <h2>Tips</h2>
           <ul>
             <li>
               Sometimes, swiping left / right on the screen triggers the Safari to go back a page. 
-              Although I cannot disable it from the website, you can disable that behavior. I'll soon link the 
-              internet guides, but you can try "Prevent swiping left to go back page Safari".
+              It's possible to disable the behavior, try "Prevent swiping left to go back page Safari".
             </li> 
           </ul>
         <br>
 
-        <h2>Roadmap</h2>
+        <!-- <h2>Other news</h2>
+
         <ul>
-          <!-- <li>Audio/video upgrades: a Zoom integration is coming</li> -->
-          <li>Blackboard upgrades: arbitrary colors, backgrounds, undo/redo, color persistence across boards etc.</li>
-          <li>Birds-eye view: for better vision of what's going on on all the boards</li>
+          <li>
+            For the life of me I cannot figure out why the ghost participants problem still remains, and I'm hoping 
+            to get second pair of eyes to take a look at my correctness argument.
+          </li>
+          <li>
+            Explain currently bleeds around $1k per month because of Twilio's video conferencing API it uses. 
+            By switching to Jitsi, an open-source alternative, the cost would become $30/month, a 3000% reduction. 
+            Previously, the switch to lib-jitsi-meet, to ensure the longevity, a major update will come and I'm working on
+            lib-jitsi-vue and testing it. 
+          </li>
+          <li>
+            I'm also going to be teaching web dev for 2021 IAP, which will be based off of lightweight fullstack (see SIPB 2020).
+            The class will take place on Explain itself, and you will work on your own projects. 
+          </li>
         </ul>
+        <br> -->
 
       </div>
       <!-- <div style="font-size: 3.75rem; font-weight: 400">
